partial update to support sending 1-255 values
It was pointed out that when working with multiple projects containing .exe of this project 0 and 1 states are not enough to determine properly inserted change in miunske. To solve this Im extending the range of values that are sent via CAN port 400 to 1-255. It creates more unique "IDs" to the separate projects. It all works by this commit and Im currently waiting for colleague to update diode module software accordingly so I can output information about expected diode behaviour